+++ src/jkind/slicing/CounterexampleSlicer.java
@@ -1,5 +1,6 @@
 package jkind.slicing;
 
+import java.math.BigInteger;
 import java.util.Map;
 import java.util.Set;
 
@@ -19,8 +20,8 @@ public class CounterexampleSlicer {
 		Model sliced = new Model();
 		for (String fn : model.getFunctions()) {
 			if (fn.startsWith("$") && keep.contains(fn.substring(1))) {
-				Map<Integer, Value> fnMap = model.getFunction(fn);
-				for (int i : fnMap.keySet()) {
+				Map<BigInteger, Value> fnMap = model.getFunction(fn);
+				for (BigInteger i : fnMap.keySet()) {
 					sliced.addFunctionValue(fn, i, fnMap.get(i));
 				}
 			}
